refactor(blocks-construct): refactor code to the convention way
