+++ src/features/profile/hooks/use-mfa.ts
@@ -1,47 +1,12 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useGlobalMutation, useGlobalQuery } from 'state/query-client/hooks';
+import { useGlobalMutation } from 'state/query-client/hooks';
 import {
-  configurationMFASave,
   generateOTP,
-  getConfigurationMFA,
   getVerifyOTP,
   manageUserMFA,
-  VerifyOTP,
 } from '../services/mfa.services';
 import { useToast } from 'hooks/use-toast';
-
-export const useGetConfigurationMfa = () => {
-  return useGlobalQuery({
-    queryKey: ['getConfigurationMFA'],
-    queryFn: getConfigurationMFA,
-  });
-};
-
-export const useSaveMfaConfiguration = () => {
-  const queryClient = useQueryClient();
-  const { toast } = useToast();
-
-  return useGlobalMutation({
-    mutationKey: ['configurationMFASave'],
-    mutationFn: configurationMFASave,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['getConfigurationMFA'] });
-      toast({
-        variant: 'success',
-        title: 'MFA Configuration Saved',
-        description: 'Multi-factor authentication has been successfully enabled for the user.',
-      });
-    },
-    onError: (error) => {
-      toast({
-        variant: 'destructive',
-        title: 'Failed to Save MFA Configuration',
-        description:
-          error?.error?.message ?? 'An error occurred while saving MFA settings. Please try again.',
-      });
-    },
-  });
-};
+import { VerifyOTP } from '../types/mfa.types';
 
 export const useGenerateOTP = () => {
   return useGlobalMutation({
