+++ src/features/profile/services/mfa.services.ts
@@ -1,48 +1,7 @@
 import { clients } from 'lib/https';
 import API_CONFIG from '../../../config/api';
+import { GenerateOTPResponse, ManageUserMFA, VerifyOTP } from '../types/mfa.types';
 
-export type ConfigMfa = {
-  enableMfa: boolean;
-  userMfaTypes: number[];
-  mFATemplate: {
-    templateName: string;
-    templateId: string;
-  };
-  projectKey: string;
-};
-
-export type ManageUserMFA = {
-  userId: string;
-  mfaEnabled: boolean;
-  userMfaType: number;
-};
-
-export type VerifyOTP = {
-  verificationCode: string;
-  twoFactorId: string;
-  authType: number;
-  projectKey: string;
-};
-
-export type GenerateOTPResponse = {
-  imageUri: string;
-  twoFactorId: string;
-  isSuccess: boolean;
-  errors?: Record<string, string>;
-};
-
-export const getConfigurationMFA = async (): Promise<ConfigMfa> => {
-  const res = await clients.get<{ data: ConfigMfa }>('/mfa/v1/Configuration/Get');
-  return res.data;
-};
-
-export const configurationMFASave = async (configMfa: ConfigMfa): Promise<ConfigMfa> => {
-  const res = await clients.post<{ data: ConfigMfa }>(
-    '/mfa/v1/Configuration/Save',
-    JSON.stringify(configMfa)
-  );
-  return res.data;
-};
 
 export const generateOTP = async ({ userId }: { userId: string }): Promise<GenerateOTPResponse> => {
   const payload = {
